+++ jfaudiolib/src/multivoc.h
@@ -79,41 +79,41 @@ int   MV_StartRecording( int MixRate, void ( *function )( char *ptr, int length 
 void  MV_StopRecord( void );
 int   MV_StartDemandFeedPlayback( void ( *function )( char **ptr, unsigned int *length ),
          int rate, int pitchoffset, int vol, int left, int right,
-         int priority, unsigned int callbackval );
+         int priority, uintptr_t callbackval );
 int   MV_PlayRaw( char *ptr, unsigned int length,
          unsigned rate, int pitchoffset, int vol, int left,
-         int right, int priority, unsigned int callbackval );
+         int right, int priority, uintptr_t callbackval );
 int   MV_PlayLoopedRaw( char *ptr, unsigned int length,
          char *loopstart, char *loopend, unsigned rate, int pitchoffset,
          int vol, int left, int right, int priority,
-         unsigned int callbackval );
+         uintptr_t callbackval );
 int   MV_PlayWAV( char *ptr, unsigned int length, int pitchoffset, int vol, int left,
-         int right, int priority, unsigned int callbackval );
+         int right, int priority, uintptr_t callbackval );
 int   MV_PlayWAV3D( char *ptr, unsigned int length, int pitchoffset, int angle, int distance,
-         int priority, unsigned int callbackval );
+         int priority, uintptr_t callbackval );
 int   MV_PlayRaw3D( char *ptr, unsigned int length, unsigned rate, int pitchoffset, int angle, int distance,
-         int priority, unsigned int callbackval );
+         int priority, uintptr_t callbackval );
 int   MV_PlayLoopedWAV( char *ptr, unsigned int length, int loopstart, int loopend,
          int pitchoffset, int vol, int left, int right, int priority,
-         unsigned int callbackval );
+         uintptr_t callbackval );
 int   MV_PlayVOC3D( char *ptr, unsigned int length, int pitchoffset, int angle, int distance,
-         int priority, unsigned int callbackval );
+         int priority, uintptr_t callbackval );
 int   MV_PlayVOC( char *ptr, unsigned int length, int pitchoffset, int vol, int left, int right,
-         int priority, unsigned int callbackval );
+         int priority, uintptr_t callbackval );
 int   MV_PlayLoopedVOC( char *ptr, unsigned int length, int loopstart, int loopend,
          int pitchoffset, int vol, int left, int right, int priority,
-         unsigned int callbackval );
+         uintptr_t callbackval );
 int   MV_PlayVorbis3D( char *ptr, unsigned int length, int pitchoffset, int angle, int distance,
-                    int priority, unsigned int callbackval );
+                    int priority, uintptr_t callbackval );
 int   MV_PlayVorbis( char *ptr, unsigned int length, int pitchoffset, int vol, int left, int right,
-                  int priority, unsigned int callbackval );
+                  int priority, uintptr_t callbackval );
 int   MV_PlayLoopedVorbis( char *ptr, unsigned int length, int loopstart, int loopend,
                         int pitchoffset, int vol, int left, int right, int priority,
-                        unsigned int callbackval );
+                        uintptr_t callbackval );
 void  MV_CreateVolumeTable( int index, int volume, int MaxVolume );
 void  MV_SetVolume( int volume );
 int   MV_GetVolume( void );
-void  MV_SetCallBack( void ( *function )( unsigned int ) );
+void  MV_SetCallBack( void ( *function )( uintptr_t ) );
 void  MV_SetReverseStereo( int setting );
 int   MV_GetReverseStereo( void );
 int   MV_Init( int soundcard, int * MixRate, int Voices, int * numchannels,
